+++ src/test/convertZipCode.test.ts
@@ -1,5 +1,5 @@
 import { expect } from '@jest/globals';
-import { goqZipCode } from './goqZipCode';
+import { goqZipCode } from './utils/goqZipCode';
 
 describe('convertZipCodeの動作をテスト', () => {
   const expectedZipCodeExcludingHyphen = '7320021';
